requirements.txt
Django==1.10.5
django-ckeditor==5.1.0
Pillow==3.3.0
python-decouple==3.0
django-imagekit==3.3
